# docs/openapi.yaml
openapi: 3.0.3
info:
  title: EntreSkill Hub API
  description: >
    Skill-to-Startup Enablement Platform API. Covers auth, assessment,
    recommendations, roadmaps, learning resources, mentors, notifications,
    and admin operations. See docs/API.md for narrative documentation.
  version: "1.0.0"
  contact:
    name: EntreSkill Hub Engineering
servers:
  - url: https://entreskillhub.com/api
    description: Production
  - url: https://staging.entreskillhub.com/api
    description: Staging
  - url: http://localhost:3000/api
    description: Local development
...
paths:
...
  /roadmaps/{id}:
    get:
      tags: [Roadmaps]
      summary: Get a roadmap and its ordered steps
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Roadmap" }
...
  /resources:
    get:
      tags: [Resources]
      summary: List approved learning resources
      security: []
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [video, article, checklist] }
        - name: tag
          in: query
          schema: { type: string }
        - name: q
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaginatedResources" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 20, maximum: 100 }
...
  schemas:
...
    Roadmap:
      type: object
      properties:
        _id: { type: string }
        businessIdeaId: { type: string }
        title: { type: string }
        steps:
          type: array
          items: { $ref: "#/components/schemas/RoadmapStep" }
    RoadmapStep:
      type: object
      properties:
        _id: { type: string }
        order: { type: integer }
        type:
          type: string
          enum: [validation, skills_tools, legal_registration, cost_estimation, marketing_basics]
        title: { type: string }
        content: { type: string }
        resourceIds: { type: array, items: { type: string } }
...
    PaginatedResources:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/LearningResource" }
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        hasMore: { type: boolean }
    LearningResource:
      type: object
      properties:
        _id: { type: string }
        title: { type: string }
        type: { type: string, enum: [video, article, checklist] }
        url: { type: string }
        uploadedBy: { type: string }
        approvalStatus: { type: string, enum: [pending, approved, rejected] }
        tags: { type: array, items: { type: string } }
...
security:
  - sessionCookie: []